backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
openai==1.3.0
PyPDF2==3.0.1
python-dotenv==1.0.0
pytest==7.4.3
pytest-cov==4.1.0
httpx==0.25.2